requirements.txt
aiohttp==3.8.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.0
async-timeout==4.0.3
asyncssh==2.13.2
atpublic==4.0
attrs==23.1.0
backcall==0.2.0
backports.functools-lru-cache==1.6.5
backports.zoneinfo==0.2.1
billiard==4.1.0
Bottleneck==1.3.5
celery==5.3.4
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cloudpickle==2.2.1
cmake==3.27.5
colorama==0.4.6
comm==0.1.4
configobj==5.0.8
contourpy==1.1.1
cryptography==41.0.4
cycler==0.11.0
cytoolz==0.12.0
dask==2023.4.1
debugpy==1.6.7
decorator==5.1.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.8.0
dpath==2.1.6
dulwich==0.21.6
dvc==3.24.0
dvc-data==2.18.1
dvc-http==2.30.2
dvc-objects==1.0.1
dvc-render==0.6.0
dvc-studio-client==0.15.0
dvc-task==0.3.0
dvclive==3.0.1
entrypoints==0.4
executing==1.2.0
filelock==3.12.4
flatten-dict==0.4.2
flufl.lock==7.1.1
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.9.1
funcy==2.0
gitdb==4.0.10
GitPython==3.1.36
grandalf==0.8
gto==1.3.0
hydra-core==1.3.2
idna==3.4
imagecodecs-lite==2019.12.3
imageio==2.31.3
importlib-metadata==6.0.0
importlib-resources==6.0.1
ipykernel==6.25.2
ipython==8.12.0
iterative-telemetry==0.0.8
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
jupyter-client==7.3.4
jupyter_core==4.12.0
kiwisolver==1.4.5
kombu==5.3.2
lazy_loader==0.3
lit==16.0.6
locket==1.0.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
natsort==8.4.0
nest-asyncio==1.5.6
networkx==3.1
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
omegaconf==2.3.0
opencv-python==4.8.1.78
orjson==3.9.7
packaging==23.1
pandas==2.0.3
parso==0.8.3
partd==1.4.0
pathlib==1.0.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pip==23.2.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydot==1.4.2
pygit2==1.13.0
Pygments==2.16.1
pygtrie==2.5.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzmq==25.1.0
requests==2.31.0
rich==13.5.3
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
scmrepo==1.3.1
semver==3.0.1
setuptools==68.0.0
shortuuid==1.0.11
shtab==1.6.4
six==1.16.0
smmap==5.0.1
sqltrie==0.8.0
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tifffile==2023.7.10
tomlkit==0.12.1
toolz==0.12.0
torch==2.0.1
torchaudio==2.0.2
torchsummary==1.5.1
torchvision==0.15.2
tornado==6.1
tqdm==4.66.1
traitlets==5.10.0
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
voluptuous==0.13.1
wcwidth==0.2.6
wheel==0.38.4
yarl==1.9.2
zc.lockfile==3.0.post1
zipp==3.17.0
